Store money as Decimal/Numeric instead of Float
Group.target_amount/contribution_amount and Transaction.amount were
Float columns, so summing many contributions (db.func.sum) accumulated
binary floating-point rounding error (0.1 + 0.1 + 0.1 != 0.3). Switched
them to Numeric(10, 2), parse incoming amounts as Decimal via
Decimal(str(value)) instead of float() to avoid inheriting float
imprecision at the input boundary, and added a JSON provider so
jsonify() serializes Decimal as a plain number for API clients.

Added an Alembic migration (batch mode, so it works on both SQLite and
Postgres) and a first real pytest suite under tests/ (previously
test_server.py was a manual script that pings the live prod URL, not
something pytest could actually run — added pytest.ini to scope
default test discovery to tests/ so it doesn't try to collect that
script too).

+++ back-end/models/group.py
@@ -8,8 +8,8 @@ class Group(db.Model):
     id = db.Column(db.String(36), primary_key=True, default=lambda: str(uuid.uuid4()))
     name = db.Column(db.String(100), nullable=False)
     description = db.Column(db.Text, nullable=True)
-    target_amount = db.Column(db.Float, nullable=False)
-    contribution_amount = db.Column(db.Float, nullable=False)
+    target_amount = db.Column(db.Numeric(10, 2), nullable=False)
+    contribution_amount = db.Column(db.Numeric(10, 2), nullable=False)
     contribution_frequency = db.Column(db.String(20), nullable=False)  # 'daily', 'weekly', 'monthly'
     max_members = db.Column(db.Integer, nullable=False)
     is_public = db.Column(db.Boolean, default=True)
